requirements.txt
flask>=1.1.2
flask_login
flask_migrate
flask_jwt_extended
flask_wtf>=0.14.3
flask_sqlalchemy>=2.5
sqlalchemy>=1.4.11
email_validator>=1.1.2
python-decouple>=3.4
gunicorn>=20.0.4
flask_caching>=1.10.1
python-dotenv
git+git://github.com/MolSSI/qcportal@v0.50a6#egg=qcportal
#qcportal